refactor: move type names into sectionTypes/
