+++ src/schema/v2/homeView/sectionTypes/HomeViewCardsSectionType.ts
@@ -7,7 +7,7 @@ import {
 } from "../../fields/pagination"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "../standardSectionFields"
 import Image from "schema/v2/image"
