+++ src/schema/v2/homeView/sectionTypes/HomeViewDiscoverMarketingCollectionType.ts
@@ -5,7 +5,7 @@ import { FeaturedLinkConnectionType } from "../../FeaturedLink/featuredLink"
 import { emptyConnection } from "../../fields/pagination"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "../standardSectionFields"
 
 export const HomeViewDiscoverMarketingCollectionType = new GraphQLObjectType<
