+++ src/schema/v2/homeView/sectionTypes/HomeViewExploreBySectionType.ts
@@ -7,7 +7,7 @@ import {
 import { ResolverContext } from "types/graphql"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "../standardSectionFields"
 import { ImageType } from "../../image"
