+++ src/schema/v2/homeView/sectionTypes/HomeViewFairsSectionType.ts
@@ -4,7 +4,7 @@ import { ResolverContext } from "types/graphql"
 import { fairsConnection } from "../../fairs"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "../standardSectionFields"
 
 export const HomeViewFairsSectionType = new GraphQLObjectType<
