+++ src/schema/v2/homeView/sectionTypes/HomeViewHeroUnitsSectionType.ts
@@ -5,7 +5,7 @@ import { emptyConnection } from "../../fields/pagination"
 import { heroUnitsConnection } from "../../HeroUnit/heroUnitsConnection"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "../standardSectionFields"
 
 export const HomeViewHeroUnitsSectionType = new GraphQLObjectType<
