+++ src/schema/v2/homeView/sectionTypes/HomeViewSalesSectionType.ts
@@ -4,7 +4,7 @@ import { ResolverContext } from "types/graphql"
 import { NodeInterface } from "../../object_identification"
 import { SalesConnectionField } from "../../sales"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "../standardSectionFields"
 
 export const HomeViewSalesSectionType = new GraphQLObjectType<
