+++ src/schema/v2/homeView/sectionTypes/HomeViewViewingRoomsSectionType.ts
@@ -2,7 +2,7 @@ import { GraphQLObjectType } from "graphql"
 import { ResolverContext } from "types/graphql"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "../standardSectionFields"
 
 export const HomeViewViewingRoomsSectionType = new GraphQLObjectType<
