+++ src/schema/v2/homeView/sectionTypes/names.ts
