+++ src/schema/v2/homeView/sections/ActiveBids.ts
@@ -1,7 +1,7 @@
 import { ContextModule } from "@artsy/cohesion"
 import { HomeViewSection } from "."
 import { withHomeViewTimeout } from "../helpers/withHomeViewTimeout"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "../sectionTypes/names"
 import { connectionFromArray } from "graphql-relay"
 
 export const ActiveBids: HomeViewSection = {
