+++ src/schema/v2/homeView/sections/ExploreByMarketingCollectionCategories.ts
@@ -1,5 +1,5 @@
 import { HomeViewSection } from "."
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "../sectionTypes/names"
 
 export const ExploreByMarketingCollectionCategories: HomeViewSection = {
   id: "home-view-section-explore-by-marketing-collection-categories",
