+++ src/schema/v2/homeView/sections/GalleriesNearYou.ts
@@ -1,6 +1,6 @@
 import { ContextModule, OwnerType } from "@artsy/cohesion"
 import { HomeViewSection } from "."
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "../sectionTypes/names"
 
 export const GalleriesNearYou: HomeViewSection = {
   id: "home-view-section-galleries-near-you",
