+++ src/schema/v2/homeView/sections/HeroUnits.ts
@@ -1,7 +1,7 @@
 import { ContextModule } from "@artsy/cohesion"
 import { HomeViewSection } from "."
 import { withHomeViewTimeout } from "../helpers/withHomeViewTimeout"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "../sectionTypes/names"
 import { heroUnitsConnection } from "schema/v2/HeroUnit/heroUnitsConnection"
 
 export const HeroUnits: HomeViewSection = {
