+++ src/schema/v2/homeView/sections/LatestAuctionResults.ts
@@ -1,7 +1,7 @@
 import { ContextModule, OwnerType } from "@artsy/cohesion"
 import { HomeViewSection } from "."
 import { withHomeViewTimeout } from "../helpers/withHomeViewTimeout"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "../sectionTypes/names"
 import {
   AuctionResultSortEnum,
   AuctionResultsStateEnums,
