+++ src/schema/v2/homeView/sections/_ExampleSection.ts
@@ -2,7 +2,7 @@ import { ContextModule, OwnerType } from "@artsy/cohesion"
 import { ResolverContext } from "types/graphql"
 import { HomeViewSection } from "."
 import { withHomeViewTimeout } from "../helpers/withHomeViewTimeout"
-import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
+import { HomeViewSectionTypeNames } from "../sectionTypes/names"
 import { filterArtworksConnectionWithParams } from "schema/v2/filterArtworksConnection"
 
 /*
